// server/@types/shared/models/RequestForPlacementStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RequestForPlacementStatus = 'request_unsubmitted' | 'request_rejected' | 'request_submitted' | 'awaiting_match' | 'request_withdrawn' | 'placement_booked';